Cleaned up formatting
