+++ gui/index.html
@@ -5,9 +5,9 @@
     <title>DAST</title>
   </head>
   <body>
-    <button id="test_button">Connect & Send</button>
+    <button id="testButton">Connect & Send</button>
     <div id="out"></div>
     
-    <script src="./test_test.js"></script>
+    <script src="./testTest.js"></script>
   </body>
 </html>
