+++ gui/testTest.js
@@ -1,15 +1,15 @@
 "use strict";
 
-document.getElementById("test_button").addEventListener("click", () => {
+document.getElementById("testButton").addEventListener("click", () => {
     const socket = new WebSocket("ws://localhost:8989");
 
     socket.addEventListener('open', () => {
         const msg = `{
-            "message-type": "CREATE_SUITE",
+            "message-type": "CREATE-SUITE",
             "url": "https://google.com",
             "tests": {
-                "test-test": {},
-                "test-test-duplicate":{}
+                "testTest": {},
+                "testTestDuplicate":{}
             }
         }`;
         socket.send(msg);
